+++ packages/patrol/android/src/main/kotlin/pl/leancode/patrol/contracts/MobileAutomatorServer.kt
@@ -39,6 +39,7 @@ interface MobileAutomatorServer {
     fun handlePermissionDialog(request: Contracts.HandlePermissionRequest)
     fun setLocationAccuracy(request: Contracts.SetLocationAccuracyRequest)
     fun setMockLocation(request: Contracts.SetMockLocationRequest)
+    fun stopMockLocation()
     fun markPatrolAppServiceReady()
     fun isVirtualDevice(): Contracts.IsVirtualDeviceResponse
     fun getOsVersion(): Contracts.GetOsVersionResponse
@@ -156,6 +157,10 @@ fun getMobileAutomatorRoutes(server: MobileAutomatorServer): RoutingHttpHandler 
     "setMockLocation" bind POST to {
       val body = json.fromJson(it.bodyString(), Contracts.SetMockLocationRequest::class.java)
       server.setMockLocation(body)
+      Response(OK)
+    },
+    "stopMockLocation" bind POST to {
+      server.stopMockLocation()
       Response(OK)
     },
     "markPatrolAppServiceReady" bind POST to {